+++ src/Products/PloneMeeting/browser/views.py
@@ -730,7 +730,7 @@ class BaseDGHV(object):
                    checkNeedSeparator=True,
                    addCSSClass=None,
                    use_safe_html=True,
-                   clean=False):
+                   clean=True):
         """Helper method to format a p_xhtmlContents.  The xhtmlContents is a list or a string containing
            either XHTML content or some specific recognized words like :
            - 'separator', in this case, it is replaced with the p_separatorValue;
@@ -2281,12 +2281,13 @@ class UpdateLocalRolesBatchActionForm(BaseBatchActionForm):
 
     def available(self):
         """Hide it on meetings as it uses IMeetingBatchActionsMarker."""
-        return _checkPermission(ManagePortal, self.context) and not IMeeting.providedBy(self.context)
+        return _checkPermission(ManagePortal, self.context) and \
+            not IMeeting.providedBy(self.context)
 
     def _apply(self, **data):
         """ """
         uids = listify_uids(data['uids'])
-        self.tool.updateAllLocalRoles(**{'UID': uids})
+        self.tool.updateAllLocalRoles(**{'UID': uids, 'log': False})
         msg = translate('update_selected_elements',
                         domain="PloneMeeting",
                         mapping={'number_of_elements': len(uids)},