+++ src/Products/PloneMeeting/utils.py
@@ -11,9 +11,6 @@ from collective.contact.plonegroup.utils import get_own_organization
 from collective.contact.plonegroup.utils import get_plone_group
 from collective.contact.plonegroup.utils import get_plone_group_id
 from collective.excelexport.exportables.dexterityfields import get_exportable_for_fieldname
-from collective.fingerpointing.config import AUDIT_MESSAGE
-from collective.fingerpointing.logger import log_info
-from collective.fingerpointing.utils import get_request_information
 from collective.iconifiedcategory.interfaces import IIconifiedInfos
 from DateTime import DateTime
 from datetime import datetime
@@ -23,6 +20,7 @@ from email.MIMEBase import MIMEBase
 from email.MIMEMultipart import MIMEMultipart
 from email.MIMEText import MIMEText
 from imio.helpers.content import richtextval
+from imio.helpers.security import fplog
 from imio.helpers.xhtml import addClassToLastChildren
 from imio.helpers.xhtml import CLASS_TO_LAST_CHILDREN_NUMBER_OF_CHARS_DEFAULT
 from imio.helpers.xhtml import markEmptyTags
@@ -1011,13 +1009,6 @@ def notifyModifiedAndReindex(obj, extra_idxs=[], notify_event=False):
 
     if notify_event:
         notify(ObjectEditedEvent(obj))
-
-
-def fplog(action, extras):
-    """collective.fingerpointing add log message."""
-    # add logging message to fingerpointing log
-    user, ip = get_request_information()
-    log_info(AUDIT_MESSAGE.format(user, ip, action, extras))
 
 
 def transformAllRichTextFields(obj, onlyField=None):